requirements.txt
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.1
click==8.1.8
cssselect2==0.8.0
defusedxml==0.7.1
et_xmlfile==2.0.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
pycairo==1.28.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
tinycss2==1.4.0
typing_extensions==4.13.2
tzdata==2025.2
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
pyinstaller
gunicorn
flask-sqlalchemy
psycopg2-binary
